+++ gateway/core/__init__.py
@@ -1,1 +1,0 @@
-"""Gateway core utilities (JWT, crypto, exceptions)."""